[Issue-1735] Add condition location for App banner, App popup
